refactor: extract xhttptest
