+++ xhttptest/doc.go
@@ -1,0 +1,2 @@
+// Package xhttptest eXtends Go's httptest.
+package xhttptest
